feat: Add Pre06, Post05 labs and update Pre05 problems
Adds the laboratory sessions for Pre06 and Post05, including raw and answered versions of the problems.

Updates Pre05 Problem 4 by replacing the complex "Secret Hitler" implementation with a more straightforward student template. This change is intended to simplify the problem for learning purposes.

In Pre05 Problem 1, the main executable has been removed and a HybridCar.java file has been added.